+++ test/unit/pullConsumers/prometheusPullConsumerTestsData.js
@@ -323,6 +323,70 @@ f5_activeHandshakeRejected{serverSslProfiles="/Common/serverssl-insecure-compati
 f5_activeHandshakeRejected{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
 f5_activeHandshakeRejected{serverSslProfiles="/Common/wom-default-serverssl"} 0
 
+# HELP f5_cipherUses_adhKeyxchg cipherUses.adhKeyxchg
+# TYPE f5_cipherUses_adhKeyxchg gauge
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_adhKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_adhKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_aesBulk cipherUses.aesBulk
+# TYPE f5_cipherUses_aesBulk gauge
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_aesBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_aesBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_aesGcmBulk cipherUses.aesGcmBulk
+# TYPE f5_cipherUses_aesGcmBulk gauge
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_aesGcmBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_aesGcmBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_camelliaBulk cipherUses.camelliaBulk
+# TYPE f5_cipherUses_camelliaBulk gauge
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_camelliaBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_camelliaBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
 # HELP f5_cipherUses_chacha20Poly1305Bulk cipherUses.chacha20Poly1305Bulk
 # TYPE f5_cipherUses_chacha20Poly1305Bulk gauge
 f5_cipherUses_chacha20Poly1305Bulk{clientSslProfiles="/Common/clientssl"} 0
@@ -338,6 +402,262 @@ f5_cipherUses_chacha20Poly1305Bulk{serverSslProfiles="/Common/serverssl"} 0
 f5_cipherUses_chacha20Poly1305Bulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
 f5_cipherUses_chacha20Poly1305Bulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
 f5_cipherUses_chacha20Poly1305Bulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_desBulk cipherUses.desBulk
+# TYPE f5_cipherUses_desBulk gauge
+f5_cipherUses_desBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_desBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_desBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_desBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_desBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_desBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_desBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_dhRsaKeyxchg cipherUses.dhRsaKeyxchg
+# TYPE f5_cipherUses_dhRsaKeyxchg gauge
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_dhRsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_dhRsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_dheDssKeyxchg cipherUses.dheDssKeyxchg
+# TYPE f5_cipherUses_dheDssKeyxchg gauge
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_dheDssKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_dheDssKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_ecdhEcdsaKeyxchg cipherUses.ecdhEcdsaKeyxchg
+# TYPE f5_cipherUses_ecdhEcdsaKeyxchg gauge
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_ecdhEcdsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_ecdhRsaKeyxchg cipherUses.ecdhRsaKeyxchg
+# TYPE f5_cipherUses_ecdhRsaKeyxchg gauge
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_ecdhRsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_ecdheEcdsaKeyxchg cipherUses.ecdheEcdsaKeyxchg
+# TYPE f5_cipherUses_ecdheEcdsaKeyxchg gauge
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_ecdheEcdsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_ecdheRsaKeyxchg cipherUses.ecdheRsaKeyxchg
+# TYPE f5_cipherUses_ecdheRsaKeyxchg gauge
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_ecdheRsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_edhRsaKeyxchg cipherUses.edhRsaKeyxchg
+# TYPE f5_cipherUses_edhRsaKeyxchg gauge
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_edhRsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_edhRsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_ideaBulk cipherUses.ideaBulk
+# TYPE f5_cipherUses_ideaBulk gauge
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_ideaBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_ideaBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_md5Digest cipherUses.md5Digest
+# TYPE f5_cipherUses_md5Digest gauge
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_md5Digest{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_md5Digest{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_nullBulk cipherUses.nullBulk
+# TYPE f5_cipherUses_nullBulk gauge
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_nullBulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_nullBulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_nullDigest cipherUses.nullDigest
+# TYPE f5_cipherUses_nullDigest gauge
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_nullDigest{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_nullDigest{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_rc2Bulk cipherUses.rc2Bulk
+# TYPE f5_cipherUses_rc2Bulk gauge
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_rc2Bulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_rc2Bulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_rc4Bulk cipherUses.rc4Bulk
+# TYPE f5_cipherUses_rc4Bulk gauge
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_rc4Bulk{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_rc4Bulk{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_rsaKeyxchg cipherUses.rsaKeyxchg
+# TYPE f5_cipherUses_rsaKeyxchg gauge
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_rsaKeyxchg{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_rsaKeyxchg{serverSslProfiles="/Common/wom-default-serverssl"} 0
+
+# HELP f5_cipherUses_shaDigest cipherUses.shaDigest
+# TYPE f5_cipherUses_shaDigest gauge
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/clientssl"} 0
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/clientssl-insecure-compatible"} 0
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/clientssl-secure"} 0
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/crypto-server-default-clientssl"} 0
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/splitsession-default-clientssl"} 0
+f5_cipherUses_shaDigest{clientSslProfiles="/Common/wom-default-clientssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/apm-default-serverssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/crypto-client-default-serverssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/pcoip-default-serverssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/serverssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/serverssl-insecure-compatible"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/splitsession-default-serverssl"} 0
+f5_cipherUses_shaDigest{serverSslProfiles="/Common/wom-default-serverssl"} 0
 
 # HELP f5_currentCompatibleConnections currentCompatibleConnections
 # TYPE f5_currentCompatibleConnections gauge